Add 4.21*: bloglist expansion, step10

+++ part4/bloglist/controllers/blogs.js
@@ -56,10 +56,20 @@ blogsRouter.put('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      return response.status(401).json({ error: 'a blog can only be deleted by the user who added the blog' })
+    }
   } catch (exception) {
+    response.status(400).end()
     next(exception)
   }
 })
